Split unit tests out of glean-core/src/lib.rs
